backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.6,<3.0
passlib[bcrypt]>=1.7,<2.0
PyJWT>=2.9,<3.0
python-multipart>=0.0.9,<1.0
requests>=2.32,<3.0
Pillow>=10.4,<11.0
livekit-api>=0.8,<1.0
pytest>=8.3,<9.0
httpx>=0.27,<1.0

